requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Jinja2==2.11.1
Werkzeug==1.0.0
Pillow
# collections
numpy
opencv-python
scikit-image
oss2